+++ ui/components/MainFooter.vue
@@ -1,5 +1,5 @@
 <template>
-    <footer class="footer">
+    <footer class="footer mt-5">
         <div class="content">
             <p class="is-size-7 has-text-weight-light has-text-centered">© {{ new Date().getFullYear() }} Rubix ML</p>
         </div>