+++ ui/pages/BubblePlot.vue
@@ -2,7 +2,7 @@
     <div>
         <section class="section">
             <div class="container">
-                <h2 class="title is-size-5 mt-5"><span class="icon mr-3"><i class="fas fa-cogs"></i></span>Chart Properties</h2>
+                <h2 class="title is-size-5"><span class="icon mr-3"><i class="fas fa-cogs"></i></span>Chart Properties</h2>
                 <div class="columns">
                     <div class="column is-one-third">
                         <div class="field">
@@ -118,6 +118,12 @@
                 </figure>
             </div>
         </section>
+        <section class="section">
+            <div class="container">
+                <h2 class="title is-size-5"><span class="icon mr-3"><i class="fas fa-file-export"></i></span>Export Chart</h2>
+                <export-chart v-if="canvas" :canvas="canvas"></export-chart>
+            </div>
+        </section>
     </div>
 </template>
 
@@ -143,6 +149,7 @@ export default Vue.extend({
                     b: 222,
                 },
             },
+            canvas: null,
             chart: null,
         };
     },
@@ -178,6 +185,8 @@ export default Vue.extend({
 
         const context = element.getContext('2d');
 
+        this.canvas = element;
+
         this.chart = new Chart(context, {
             type: 'bubble',
             data: {
@@ -194,8 +203,8 @@ export default Vue.extend({
                 responsive: true,
                 maintainAspectRatio: false,
                 title: {
-                    display: true,
-                    text: 'Bubble Chart',
+                    text: 'Bubble Plot',
+                    display: false,
                 },
                 legend: {
                     display: false,