+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+- 0.2.6-beta
+    - Added export chart feature
+
 - 0.2.5-beta
     - Added Backoff and Retry client middleware
     - Added server memory Circuit Breaker internal middleware